Create a PetForm component and change the form for adding a new pet to a popup version
