+++ src/components/Header/Header.module.css
@@ -1,6 +1,6 @@
 .headerContainer { @apply
     relative
-    z-50
+    z-0
     flex
     mx-auto
     xl:w-[1280px];
