+++ src/styles/account/index.module.css
@@ -91,4 +91,5 @@
     gap-x-6 
     gap-y-6
     grid-cols-12
+    mb-10
 }
